refactor: rename Eacosystem Actors index path
