+++ src/views/EcosystemActorsIndexView/components/ActorCustomItem/CustomItemAll.tsx
