+++ src/views/EcosystemActorsIndexView/components/ActorCustomItem/CustomItemScope.tsx
