+++ src/views/EcosystemActorsIndexView/components/ActorItem/ActorItem.tsx
