+++ src/views/EcosystemActorsIndexView/components/ActorItem/GroupedScopes.tsx
