+++ src/views/EcosystemActorsIndexView/utils/utils.ts
